fetch!: user can change title and date with click on the element
